+++ src/util/CSSParsing/CSSTextTransformationMap.ts
@@ -1,5 +1,5 @@
-import type { Text } from '../../../shapes/Text/Text';
-import { parseUnit } from '../svgParsing';
+import type { Text } from '../../shapes/Text/Text';
+import { parseUnit } from '../misc/svgParsing';
 import { CSSTransformationMap } from './CSSTransformationMap';
 import { CSSTransformConfig, CSSTransformConfigMap } from './types';
 import {
